* origin/2.x: (104 commits)
  Change the RSS feed URLs to HTTPS
  MODX Revolution 2.7.1-pl
  Update lexicons from crowdin
  Change after review
  Update phpThumb 1.7.15-201902101903
  Restore html in resource tree (#14358) while preserving XSS protections in trees by default
  Handle deprecated $type and $responseCode parameters in $modx->sendRedirect and fix message
  Update lexicon entry
  Include not deleted children of deleted parents in the list
  Using cltr/cmd and click will open the url in a new tab/window again for ExtJS elements that use `loadPage()` to open URLs
  Forbid generating child resources for deleted resources
  Fix #14094
  XSS in the tree
  Fix #14105
  Fix #14104
  Fix #14103
  Fix #14102
  Enable remote avatars
  Fix regression in resourcelist that prevents parents from working correctly
  Improve wording in variables
  ...

+++ core/model/modx/modelement.class.php
@@ -188,7 +188,7 @@ class modElement extends modAccessibleSimpleObject {
         }
 
         /* Removing old static file when succesfull saved and oldPath has been set. */
-        if ($saved && $oldPath) {
+        if ($saved && $oldPath && $this->isStaticFilesAutomated()) {
             if (@unlink($oldPath)) {
                 $pathinfo = pathinfo($oldPath);
                 $this->cleanupStaticFileDirectories($pathinfo['dirname']);
@@ -196,6 +196,28 @@ class modElement extends modAccessibleSimpleObject {
         }
 
         return $saved;
+    }
+
+    /**
+     * Determine if static files should be automated for current element class.
+     *
+     * @return bool
+     */
+    protected function isStaticFilesAutomated()
+    {
+        $elements = array(
+            'modTemplate'    => 'templates',
+            'modTemplateVar' => 'tvs',
+            'modChunk'       => 'chunks',
+            'modSnippet'     => 'snippets',
+            'modPlugin'      => 'plugins'
+        );
+
+        if (!array_key_exists($this->_class, $elements)) {
+            return false;
+        }
+
+        return (bool) $this->xpdo->getOption('static_elements_automate_' . $elements[$this->_class], null, false);
     }
 
     /**
